More fixes for to get the “follow” link working
I thought I’d dealt with this in #53 but apparently not. To get the “follow” link in the avatar dropdown working properly in the widgets on the dashboard pages, I needed to make sure the custom.js file loads in the footer and add the “data” pieces to <div class="update-item" data-bp-item-id="<?php echo $person; ?>" data-bp-item-component="members"> - line 29 in widget.php.